discord-memo-bot/requirements.txt
discord.py>=2.4.0
requests>=2.31.0

